backend/requirements.txt
# Core
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
python-dotenv==1.1.0
gunicorn==23.0.0

# ML / Data
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
joblib==1.4.2
matplotlib==3.10.1
seaborn==0.13.2
shap==0.46.0
torch
torchvision

# Computer Vision
opencv-python-headless==4.11.0.86

# Speech / Audio
SpeechRecognition==3.14.1
librosa==0.10.2.post1
soundfile==0.13.1

# Utilities
werkzeug==3.1.3
tqdm==4.67.1
PyJWT==2.8.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
